// src/app/modules/models/UserMst.ts/UserMstDto.ts
export interface UserMstDto {
        tenentId: number;
        userId: number;
        locationId: number;
        crupId: number;
        firstName: string;
        lastName: string;
        firstName1: string;
        lastName1: string;
        firstName2: string;
        lastName2: string;
        loginId: string;
        password: string;
        userType: number;
        remarks: string;
        activeFlag: string;
        lastLoginDt: Date;
        userDetailId: number;
        accLock: string;
        firstTime: string;
        passwordChng: string;
        themeName: string;
        approvalDt: Date;
        verificationCd: string;
        emailAddress: string;
        tillDt: Date;
        isSignature: boolean;
        signatureImage: string;
        avtar: string;
        compId: number;
        empId: number;
        checkinSwitch: boolean;
        loginActive: number;
        activeuser: boolean;
        userdate: Date;
        language1: string;
        language2: string;
        language3: string;
        dateOfBirth: Date;
        employeePosition: string;
        salary: string;
        dateOfJoining: Date; 
}